Fixed agni logs
