+++ input/rce_mixed.cfg
@@ -77,8 +77,8 @@ F_crit                      = 0.2       # Model will terminate when |F_atm| < F_
 
 # Method for solving for T(p) profile
 atmosphere_model            = 1         # Atmosphere model to be used, 0: JANUS | 1: AGNI
-atmosphere_solve_energy     = 1         # Enable time-stepped atmosphere solution, 0: disabled | 1: enabled 
-atmosphere_surf_state       = 2         # Atmosphere bottom edge boundary condition, 0: free | 1: fixed at T_surf | 2: conductive skin
+atmosphere_solve_energy     = 0         # Enable time-stepped atmosphere solution, 0: disabled | 1: enabled 
+atmosphere_surf_state       = 1         # Atmosphere bottom edge boundary condition, 0: free | 1: fixed at T_surf | 2: conductive skin
 
 # Number of levels
 atmosphere_nlev             = 45
