+++ utils/helper.py
@@ -120,33 +120,6 @@ def CleanDir(directory, keep_stdlog=False):
     else:
         os.makedirs(directory)
 
-# Fake file-like stream object that redirects writes to a logger instance.
-class StreamToLogger(object):
-    # https://stackoverflow.com/a/36296215
-    def __init__(self, logger, log_level=logging.INFO):
-        self.logger = logger
-        self.log_level = log_level
-        self.linebuf = ''
-
-    def write(self, buf):
-        temp_linebuf = self.linebuf + buf
-        self.linebuf = ''
-        for line in temp_linebuf.splitlines(True):
-            # From the io.TextIOWrapper docs:
-            #   On output, if newline is None, any '\n' characters written
-            #   are translated to the system default line separator.
-            # By default sys.stdout.write() expects '\n' newlines and then
-            # translates them so this is still cross platform.
-            if line[-1] == '\n':
-                self.logger.log(self.log_level, line.rstrip())
-            else:
-                self.linebuf += line
-
-    def flush(self):
-        if self.linebuf != '':
-            self.logger.log(self.log_level, self.linebuf.rstrip())
-        self.linebuf = ''
-
 
 def find_nearest(array, target):
     """Find the element of an array that has a value closest to the target
